# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tourist",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "localhost",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
...
        "/usuarios/list": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listado de usuarios",
                "operationId": "App\\Http\\Controllers\\API\\UsuariosController::list_usuarios",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/usuarios/add": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Adiciona un usuario",
                "operationId": "App\\Http\\Controllers\\API\\UsuariosController::add_usuarios",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "tourist"
                                        ]
                                    },
                                    "nombres": {
                                        "type": "string"
                                    },
                                    "apellido1": {
                                        "type": "string"
                                    },
                                    "apellido2": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@gmail.com",
                                    "rol": "tourist",
                                    "nombres": "Pepe",
                                    "apellido1": "de la Paz",
                                    "apellido2": "Garcia"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/usuarios/update/{idusuario}": {
            "patch": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Modificar un usuario",
                "operationId": "App\\Http\\Controllers\\API\\UsuariosController::update_usuarios",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "Identificador del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "tourist"
                                        ]
                                    },
                                    "nombres": {
                                        "type": "string"
                                    },
                                    "apellido1": {
                                        "type": "string"
                                    },
                                    "apellido2": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@gmail.com",
                                    "rol": "tourist",
                                    "nombres": "Pepe",
                                    "apellido1": "de la Paz",
                                    "apellido2": "Garcia"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/usuarios/delete/{idusuario}": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "App\\Http\\Controllers\\API\\UsuariosController::delete_usuarios",
                "parameters": [
                    {
                        "name": "idusuario",
                        "in": "path",
                        "description": "Identificador del usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "login",
                "name": "token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}